pyproject.toml
[tool.poetry]
name = "menu"
version = "0.1.0"
description = "y_lab_test"
authors = ["Игорь Арефьев <djifrost@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.5"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
gunicorn = "^21.2.0"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
yarl = "^1.9.4"
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^7.0.0"


[tool.poetry.group.test.dependencies]
pytest = "7.4.4"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
